requirements.txt
# Post-Quantum Cryptography
pqcrypto==0.4.0
pycryptodomex==3.20.0
liboqs-python==0.14.1

# Cryptography
cryptography==41.0.7

# Web Framework
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.6
python-socketio==5.11.1

# JSON Web Tokens
pyjwt==2.8.0

# Utilities
requests==2.31.0
python-dateutil==2.8.2
click==8.1.7

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Development
black==23.12.1
flake8==7.0.0

# Benchmarking / Data
matplotlib==3.8.2
numpy==1.26.2
pandas==2.1.4
psutil==5.9.8

# Rust extension build tooling
maturin>=1.5,<2.0